refactor: 패키지 정리 및 우승자 이름출력 수정
