+++ src/main/java/racingCar/domain/Name.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.Objects;
