+++ src/main/java/racingCar/domain/RandomNumber.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.Random;
