+++ src/test/java/racingCar/CarTest.java
@@ -1,9 +1,9 @@
 package racingCar;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingCar.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
