+++ src/test/java/racingCar/NameTest.java
@@ -1,8 +1,8 @@
 package racingCar;
 
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingCar.domain.Name;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
